+++ src/steps/AddCardStepper.tsx
@@ -1,6 +1,6 @@
-import AddCardForm from 'src/steps/AddCardForm.tsx';
-import CardList from 'src/steps/CardList.tsx';
-import AddCardFinish from 'src/steps/AddCardFinish.tsx';
+import AddCardForm from 'src/steps/add-card-form/AddCardForm.tsx';
+import CardList from 'src/steps/card-list/CardList.tsx';
+import AddCardFinish from 'src/steps/add-card-finish/AddCardFinish.tsx';
 import { AddCardMachineProvider } from 'src/machines/addCardMachine.ts';
 
 export default function AddCardStepper() {
