+++ src/steps/add-card-form/CardExpirationDateInput.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import CardExpirationDateInputComponent from 'src/components/add-card-form/CardExpirationDateInput.tsx';
+import CardExpirationDateInputComponent from 'src/steps/add-card-form/CardExpirationDateInput.tsx';
 import { SelectToFormLayer } from 'src/components/utils/SelectToFormLayer.tsx';
 import { AddCardMachineDecorator } from 'src/stories/Decorators.tsx';
