+++ src/steps/add-card-form/CardExpirationDateInput.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import CardExpirationDateInput from 'src/components/add-card-form/CardExpirationDateInput.tsx';
+import CardExpirationDateInput from 'src/steps/add-card-form/CardExpirationDateInput.tsx';
 import { SelectToFormLayer } from 'src/components/utils/SelectToFormLayer.tsx';
 import { renderWithAddCardMachineProvider } from 'src/utils/render.tsx';
