+++ src/steps/add-card-form/CardNumberInput.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import CardNumberInput from 'src/components/add-card-form/CardNumberInput.tsx';
+import CardNumberInput from 'src/steps/add-card-form/CardNumberInput.tsx';
 import { SelectToFormLayer } from 'src/components/utils/SelectToFormLayer.tsx';
 import { renderWithAddCardMachineProvider } from 'src/utils/render.tsx';
