+++ src/steps/add-card-form/CardOwnerNameInput.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import CardOwnerNameInput from 'src/components/add-card-form/CardOwnerNameInput.tsx';
+import CardOwnerNameInput from 'src/steps/add-card-form/CardOwnerNameInput.tsx';
 import { SelectToFormLayer } from 'src/components/utils/SelectToFormLayer.tsx';
 import { renderWithAddCardMachineProvider } from 'src/utils/render.tsx';
