+++ src/steps/add-card-form/CardPasswordInput.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import CardPasswordInput from 'src/components/add-card-form/CardPasswordInput.tsx';
+import CardPasswordInput from 'src/steps/add-card-form/CardPasswordInput.tsx';
 import { SelectToFormLayer } from 'src/components/utils/SelectToFormLayer.tsx';
 import { renderWithAddCardMachineProvider } from 'src/utils/render.tsx';
