+++ src/steps/add-card-form/CardSecurityCodeInput.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import CardSecurityCodeInputComponent from 'src/components/add-card-form/CardSecurityCodeInput.tsx';
+import CardSecurityCodeInputComponent from 'src/steps/add-card-form/CardSecurityCodeInput.tsx';
 import { SelectToFormLayer } from 'src/components/utils/SelectToFormLayer.tsx';
 import { AddCardMachineDecorator } from 'src/stories/Decorators.tsx';
