+++ src/steps/add-card-form/CardSecurityCodeInput.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import CardSecurityCodeInput from 'src/components/add-card-form/CardSecurityCodeInput.tsx';
+import CardSecurityCodeInput from 'src/steps/add-card-form/CardSecurityCodeInput.tsx';
 import { SelectToFormLayer } from 'src/components/utils/SelectToFormLayer.tsx';
 import { renderWithAddCardMachineProvider } from 'src/utils/render.tsx';
